Refactor #319 - `collection ex` to `acc ex`
